feat: remove unused code, implement the choice of a random city on the login view
