+++ project/src/components/app/app.tsx
@@ -1,7 +1,7 @@
 import browserHistory from '../../browser-history';
 import Favorites from '../views/favorites/favorites';
 import Loading from '../common/loading/loading';
-import Login from '../views/login';
+import Login from '../views/login/login';
 import Main from '../views/main/main';
 import NotFound from '../views/not-found/not-found';
 import PrivateRoute from './private-route';
