+++ project/src/types/action.ts
@@ -1,4 +1,4 @@
-import { Action, ThunkAction, ThunkDispatch } from '@reduxjs/toolkit';
+import { Action, ThunkAction } from '@reduxjs/toolkit';
 import { AxiosInstance } from 'axios';
 import { State } from './store';
 
@@ -22,5 +22,3 @@ export enum ActionType {
 }
 
 export type ThunkActionResult<R = Promise<void>> = ThunkAction<R, State, AxiosInstance, Action>;
-
-export type ThunkAppDispatch = ThunkDispatch<State, AxiosInstance, Action>;
